[v.1.9.3.4] Security fix ready to release

+++ mod_news_pro_gk5/data_sources/com_content/model.php
@@ -192,7 +192,7 @@ class NSP_GK5_com_content_Model {
 			JRequest::getCmd('view') == 'article' &&
 			JRequest::getVar('id') != ''
 		) {
-			$id = JRequest::getVar('id');
+			$id = (int) JRequest::getVar('id');
 			// filter the alias from ID
 			if(stripos($id, ':') !== FALSE) {
 				$id = explode(':', $id);
